+++ frontend/src/components/chat-page.tsx
@@ -278,12 +278,18 @@ export function ChatPage({
             onKeyDown={(e) => {
               if (e.key === "Enter" && !e.shiftKey) {
                 e.preventDefault();
-                setIsAutoScrollEnabled(true);
-                handleSubmit(e);
+                if (
+                  !(
+                    isLoading ||
+                    (messages.at(-1)?.toolInvocations ?? []).length > 0
+                  )
+                ) {
+                  setIsAutoScrollEnabled(true);
+                  handleSubmit(e);
+                }
               }
             }}
             autoComplete="off"
-            disabled={isLoading}
           />
           <div className="flex gap-2 mr-3">
             <ToolSelectionDropdown
@@ -294,7 +300,12 @@ export function ChatPage({
             {isLoading ? (
               <div className="w-6 h-6 border-2 ml-2 p-1 border-gray-500 border-t-transparent rounded-full animate-spin" />
             ) : (
-              <button type="submit" data-testid="send-button" className="p-1">
+              <button
+                type="submit"
+                data-testid="send-button"
+                className="p-1"
+                disabled={(messages.at(-1)?.toolInvocations ?? []).length > 0}
+              >
                 <Send className="opacity-50" />
               </button>
             )}