+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Copyright and acknowledgement.
 - pre-commit hooks for frontend.
 - [frontend] - Loading bar and better first message.
+- Better system prompt
 
 ### Added
 - Docker compose for local development
@@ -51,6 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [frontend] - Possibility to dynamically chose available tools.
 - [frontend] - Search bar in tool list.
 - Displaying tool metadata
+- GET threads depends on vlab+project_id.
+- Object storage + plotting logic
 
 ### Fixed
 - [frontend] - Parallel execution of tools
@@ -66,8 +69,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Do not return entire coordinates in `scs-get-one` tool.
 - [frontend] - Better Vlab and Project selection.
 - [frontend] - Fix long message overflowing.
+- Display the entire input json for tool calls.
+- Push to ECR prod+Staging.
+- [frontend] - Text input disabled when HIL.
+- Tool + content is saved as `AI_tool` now.
 - [frontend] - Scrolling while streaming.
+- [frontend] - Handle invalid json errors.
 - [frontend] - Tool selection is all checked when refreshing.
+- Fix bug for AI + tool calls.
 
 ## [0.3.3] - 30.10.2024